pyproject.toml
[project]
name = "automation-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "docling>=2.74.0",
    "gitingest>=0.3.1",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.10",
    "langchain-core>=1.2.15",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langsmith>=0.7.6",
    "openai>=2.23.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
